backend/soap_service/requirements.txt
fastapi
uvicorn
lxml
psycopg2-binary
